+++ app/charts/combo/combo-line-dual-state.tsx
@@ -37,6 +37,7 @@ import { ComboLineDualConfig } from "@/configurator";
 import { Observation } from "@/domain/data";
 import { truthy } from "@/domain/types";
 import { getTextWidth } from "@/utils/get-text-width";
+import { useAxisTitleAdjustments } from "@/utils/use-axis-title-adjustments";
 import { useIsMobile } from "@/utils/use-is-mobile";
 
 import { ChartProps } from "../shared/ChartProps";
@@ -133,23 +134,11 @@ const useComboLineDualState = (
     )
   );
 
-  const axisTitleLeft = variables.y.left.label;
-  const axisTitleRight = variables.y.right.label;
-  const axisTitleWidthLeft =
-    getTextWidth(axisTitleLeft, { fontSize: TICK_FONT_SIZE }) + TICK_PADDING;
-  const axisTitleWidthRight =
-    getTextWidth(axisTitleRight, { fontSize: TICK_FONT_SIZE }) + TICK_PADDING;
-  const overLappingTitles = axisTitleWidthLeft + axisTitleWidthRight > width;
-  const overLappingAmount = (axisTitleWidthLeft + axisTitleWidthRight) / width;
-
-  const axisTitleAdjustment =
-    (overLappingTitles
-      ? TICK_FONT_SIZE * Math.ceil(overLappingAmount)
-      : TICK_FONT_SIZE + TITLE_VPADDING) *
-      2 -
-    TICK_FONT_SIZE * 2;
-
-  const topMarginAxisTitleAdjustment = 50 + axisTitleAdjustment;
+  const { topMarginAxisTitleAdjustment } = useAxisTitleAdjustments({
+    leftAxisTitle: variables.y.left.label,
+    rightAxisTitle: variables.y.right.label,
+    containerWidth: width,
+  });
 
   const right = Math.max(maxRightTickWidth, 40);