+++ app/utils/animation-frame.ts
@@ -1,0 +1,3 @@
+export const animationFrame = () => {
+  return new Promise((resolve) => requestAnimationFrame(resolve));
+};