+++ CHANGELOG.md
@@ -11,7 +11,8 @@ You can also check the
 
 # Unreleased
 
-Nothing yet.
+- Features
+  - It's now possible to export charts as images
 
 # [5.0.2] - 2024-11-28
 
@@ -22,6 +23,7 @@ Nothing yet.
   - Changed component id separator text to be more unique, as it was causing
     issues with some cubes that had iris ending with "-"
   - Datasets section in the sidebar is now correctly shown at all times
+  - Improved Preventation of overlapping axis titles for combo charts
 
 # [5.0.1] - 2024-11-26